% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data}
\alias{get_data}
\title{Retrieve data from the Eurostat API}
\usage{
get_data(key, filter = NULL)
}
\arguments{
\item{key}{A character string identifying the series to be retrieved}

\item{filter}{A named list with additional filters (see \code{details})}
}
\value{
A data frame
}
\description{
Retrieve data from the Eurostat API
}
\details{
The \code{filter} option of \code{get_data()} takes a named list of key-value
pairs. If left blank, it returns all data for the current version.

Available filter parameters:

\itemize{
\item \code{startPeriod} & \code{endPeriod}
 \itemize{
   \item \code{YYYY} for annual data (e.g.: 2013)
   \item \code{YYYY-S[1-2]} for semi-annual data (e.g.: 2013-S1)
   \item \code{YYYY-Q[1-4]} for quarterly data (e.g.: 2013-Q1)
   \item \code{YYYY-MM} for monthly data (e.g.: 2013-01)
   \item \code{YYYY-W[01-53]} for weekly data (e.g.: 2013-W01)
   \item \code{YYYY-MM-DD} for daily data (e.g.: 2013-01-01)
   }
 }
}
\examples{
# Get data on doctorates in US and DE
df <- get_data("cdh_e_occ1.A.TOTAL.PC..DE+US")
head(df)
}

