% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Retrieve dimensions of Eurostat series}
\usage{
get_dimensions(id)
}
\arguments{
\item{id}{A character string identifying the series to be retrieved}
}
\value{
A list of data frames, one for each series retrieved
}
\description{
Retrieve dimensions of Eurostat series
}
\examples{
dims <- get_dimensions("cdh_e_occ1")
str(dims)
}

