% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse_variable_names}
\alias{parse_variable_names}
\title{Parse names of observation variables}
\usage{
parse_variable_names(x)
}
\arguments{
\item{x}{A string from the header of the first column.}
}
\value{
A list of length two: a vector of the first set of
observation variables, and the last observation variable.
}
\description{
Split out the names from Eurostat's first column header
}

