% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.eurostat}
\alias{read.eurostat}
\title{Load a dataset from Eurostat.}
\usage{
read.eurostat(file, value.name = "value")
}
\arguments{
\item{file}{Path to the file to import.}

\item{value.name}{The name to be given, in the resulting data frame, to the
variable given by this dataset.}
}
\value{
A data frame in "tidy" format.
}
\description{
Read a Eurostat data file and reformat as a data frame in "tidy" format.
}
\details{
A Eurostat data table is a tab-separated-values file, having the structure
of a cross-tab. There is a single dimension for the column headings and
possibly multiple dimensions for the row headings, as follows:--
\tabular{lrrrrr}{
 AA,BB,CC\\XX \tab x1 \tab x2 \tab x3 \tab x4 \tab ... \cr
  a1,b1,c1    \tab  v \tab  v \tab  v \tab  v \tab ... \cr
  a1,b1,c2    \tab  v \tab  v \tab  v \tab  v \tab ... \cr
  ...         \tab .. \tab .. \tab .. \tab .. \tab ..
}

The \code{AA}, \code{BB}, \code{CC}, \code{XX} are the names of the
variables that specify an observation; the \code{an}, \code{bn}, \code{cn},
\code{xn}, ... are the corresponding values of each observation
variable. The \code{v} are the values of the measured variable, whose name
is not given by Eurostat (this name can be supplied by the parameter
\code{value.name}).

The values, \code{v}, may be either a single field or a field followed by a space
followed by zero or more characters (indicating flags).

The header of the first column, \code{AA,BB,CC\\XX} contains the names of
\emph{all} the variables. If there are \eqn{N} variables, then this string
has \eqn{N-1} names separated by commas, a backslash, and one more name.

It is not clear to me what happens in the case \eqn{N < 2}.

Missing values are denoted in Eurostat data tables by the character \code{:}
}
\examples{
\dontrun{nrg <- read.eurostat("nrg_100a.tsv", value.name = "energy")}
}

