% Generated by roxygen2 (4.0.1): do not edit by hand
\name{split_and_rename_variables}
\alias{split_and_rename_variables}
\title{Split observation variables into separate columns}
\usage{
split_and_rename_variables(df)
}
\arguments{
\item{df}{A dataframe, having three columns
\enumerate{
\item The observation variables (as one column);
\item The measure variable;
\item Any flags.
}}
}
\value{
A data frame.
}
\description{
Breaks out first column of Eurostat data table into multiple variables.
}

