% Generated by roxygen2 (4.0.1): do not edit by hand
\name{split_flags}
\alias{split_flags}
\title{Split flags from the measure variable}
\usage{
split_flags(df, value.name)
}
\arguments{
\item{df}{A data frame}

\item{value.name}{The column containing the measure variable.}
}
\value{
A new data frame, with an additional column.
}
\description{
Parse the measure variable to extract any flags.
}

